backend/requirements.txt
# 🚀 Ottawa GenAI Research Assistant - Python Dependencies

# 核心Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP客户端和工具
requests>=2.31.0
httpx>=0.25.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# PDF处理
pypdf>=6.0.0
pdfplumber>=0.9.0

# 认证和安全 - 修复bcrypt版本兼容问题
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0

# PostgreSQL 数据库驱动（生产环境）
psycopg2-binary>=2.9.0

# 环境配置
python-dotenv>=1.0.0

# 模板引擎
jinja2>=3.1.0

# 日志
loguru>=0.7.0

# 邮件验证和处理
email-validator>=2.0.0

# 文件处理
aiofiles>=23.0.0

# 时间处理
python-dateutil>=2.8.0

# JSON处理增强
orjson>=3.9.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# 测试工具
httpx>=0.25.0  # 用于测试 API 端点
factory-boy>=3.3.0  # 用于创建测试数据

# Google OAuth支持
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# AI/ML相关 - 免费AI提供商
google-generativeai>=0.3.0  # Google Gemini AI (免费)
groq>=0.4.0                  # Groq AI (免费超快速)

# AI/ML相关（可选）
openai>=1.0.0
anthropic>=0.7.0

# 向量数据库（可选）
chromadb>=0.4.0
pinecone-client>=2.2.0

# 文档和字符串处理
python-slugify>=8.0.0
unidecode>=1.3.0

# 缓存
redis>=5.0.0
hiredis>=2.2.0

# 监控和性能
prometheus-client>=0.17.0 